backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
django-filter==2.4.0
djoser==2.2.2
drf-extra-fields==3.7.0
psycopg2-binary==2.9.3
Pillow==9.3.0
django-colorfield==0.11.0
python-dotenv==1.0.1
gunicorn==20.1.0